vpdo_dev_t: int port instead of ULONG

+++ driver/vhci/vhub.h
@@ -4,7 +4,7 @@
 #include "dev.h"
 #include "usbip_vhci_api.h"
 
-PAGEABLE vpdo_dev_t *vhub_find_vpdo(vhub_dev_t *vhub, ULONG port);
+PAGEABLE vpdo_dev_t *vhub_find_vpdo(vhub_dev_t *vhub, int port);
 
 PAGEABLE void vhub_attach_vpdo(vhub_dev_t *vhub, vpdo_dev_t *vpdo);
 PAGEABLE void vhub_detach_vpdo_and_release_port(vhub_dev_t *vhub, vpdo_dev_t *vpdo);
